+++ crates/zap-language-server/src/hovers/types.rs
@@ -6,35 +6,47 @@ use async_language_server::{
 };
 use zap_language::docs::find_primitive;
 
-use crate::utils::{find_declared_type, is_type, is_type_reference};
+use crate::structs::ReferencedType;
 
 pub fn hover(doc: &Document, _pos: Position, node: Node) -> Option<Hover> {
-    if !is_type(node) {
-        return None;
-    }
+    if matches!(node.kind(), "primitive_type") {
+        // Primitive type such as `u32`, `string`, etc
+        let text = doc.text().byte_slice(node.byte_range());
+        let (_, header, desc) = find_primitive([text])?;
 
-    let text = doc.text().byte_slice(node.byte_range());
-
-    if let Some((_, header, desc)) = find_primitive([text]) {
-        return Some(Hover {
+        Some(Hover {
             range: Some(ts_range_to_lsp_range(node.range())),
             contents: HoverContents::Scalar(MarkedString::String(format!(
                 "# {header}\n\n{desc}\n"
             ))),
-        });
-    }
+        })
+    } else {
+        // May be a referenced type that needs to be resolved,
+        // if we are hovering over a qualified / namespaced type
+        // we should also make sure to resolve the *full* reference
+        let node = match node.parent() {
+            Some(p) if p.kind() == "namespaced_type" => p,
+            _ => node,
+        };
 
-    if is_type_reference(node) {
-        let type_decl = find_declared_type(doc, text)?;
-        let type_contents = doc.text().byte_slice(type_decl.byte_range());
+        let typ = ReferencedType::from_node(doc, node)?;
+        let decl = typ.resolve_declaration()?;
 
-        return Some(Hover {
-            range: Some(ts_range_to_lsp_range(node.range())),
+        // We show an auto-formatted version of the type declaration
+        // here to automatically de-indent and make it easier to read
+        let text = doc.text_bytes();
+        let config = zap_formatter::Config::new(text.as_slice());
+
+        let mut formatted = String::new();
+        if zap_formatter::format_root(&mut formatted, config, *decl.as_ref()).is_err() {
+            formatted = decl.declaration_text();
+        }
+
+        Some(Hover {
+            range: Some(ts_range_to_lsp_range(typ.reference_range())),
             contents: HoverContents::Scalar(MarkedString::String(format!(
-                "```zap\n{type_contents}\n```"
+                "```zap\n{formatted}\n```",
             ))),
-        });
+        })
     }
-
-    None
 }