+++ crates/zap-language-server/src/utils.rs
@@ -1,80 +1,10 @@
-#![allow(clippy::needless_pass_by_value)]
-
-use std::collections::HashMap;
-
-use async_language_server::{server::Document, tree_sitter::Node, tree_sitter_utils::find_child};
-use zap_language::tree_sitter_utils::DepthFirstNodeIterator;
+use async_language_server::{tree_sitter::Node, tree_sitter_utils::find_child};
 
 /**
-    Checks if the given node is a declared type, meaning
-    it is the identifier inside a `type_declaration` node.
+    Checks if the given node is a type, primitive or reference.
 */
-pub fn is_declared_type(node: Node) -> bool {
-    node.parent().is_some_and(|parent| {
-        if parent.kind() == "type_declaration" {
-            parent
-                .child_by_field_name("name")
-                .is_some_and(|n| n == node)
-        } else {
-            false
-        }
-    })
-}
-
-/**
-    Finds a specific declared type by its name in the given document,
-    and returns the corresponding node for the full type declaration.
-*/
-pub fn find_declared_type(doc: &Document, type_name: impl ToString) -> Option<Node> {
-    let root = doc.node_at_root()?;
-    let name = type_name.to_string();
-
-    let mut cursor = root.walk();
-    for top_level in root.children(&mut cursor) {
-        if top_level.kind() == "type_declaration" {
-            let mut top_level_cursor = top_level.walk();
-            for child in top_level.children(&mut top_level_cursor) {
-                if child.kind() == "identifier" {
-                    let text = doc.text().byte_slice(child.byte_range());
-                    if text.as_str().is_some_and(|t| t == name.as_str()) {
-                        return Some(top_level);
-                    }
-                }
-            }
-        }
-    }
-
-    None
-}
-
-/**
-    Gathers all of the declared types in the given document, and returns
-    them all as a `HashMap` of type names to their corresponding nodes.
-*/
-pub fn gather_declared_types(doc: &Document) -> HashMap<String, Node> {
-    let Some(root) = doc.node_at_root() else {
-        return HashMap::new();
-    };
-
-    let mut cursor = root.walk();
-    let mut types = HashMap::new();
-
-    for top_level in root.children(&mut cursor) {
-        if top_level.kind() == "type_declaration" {
-            let mut top_level_cursor = top_level.walk();
-            for child in top_level.children(&mut top_level_cursor) {
-                if child.kind() == "identifier" {
-                    let text = doc.text().byte_slice(child.byte_range());
-                    if let Some(text) = text.as_str() {
-                        types.insert(text.to_string(), top_level);
-                    }
-                    break;
-                }
-            }
-        }
-    }
-
-    types
+pub fn is_type(node: Node) -> bool {
+    is_type_primitive(node) || is_type_reference(node)
 }
 
 /**
@@ -85,14 +15,7 @@ pub fn is_type_primitive(node: Node) -> bool {
 }
 
 /**
-    Checks if the given node is a type, primitive or reference.
-*/
-pub fn is_type(node: Node) -> bool {
-    is_type_primitive(node) || is_type_reference(node)
-}
-
-/**
-    Checks if the given node is a type reference.
+    Checks if the given identifier node is a type reference.
 
     Note that this does not check if the node is a **valid** type
     reference, only that it is *trying* to reference some type.
@@ -108,8 +31,8 @@ pub fn is_type_reference(node: Node) -> bool {
     };
 
     match p.kind() {
-        // Optionals, properties, and sets have an inner "type" field
-        "optional_type" | "property" | "set_type" => p
+        // These all have an inner "type" field
+        "namespaced_type" | "optional_type" | "property" | "set_type" => p
             .child_by_field_name("type")
             .is_some_and(is_ident_and_matches_this_node),
 
@@ -139,15 +62,4 @@ pub fn is_type_reference(node: Node) -> bool {
         // Nothing else can be a type reference according to grammar
         _ => false,
     }
-}
-
-/**
-    Gathers all of the type references in the given node, searching
-    all descendants and returning them in depth-first node order.
-*/
-pub fn gather_type_references(node: Node) -> Vec<Node> {
-    DepthFirstNodeIterator::new(node)
-        .filter(|n| n.kind() == "identifier")
-        .filter(|n| is_type_reference(*n))
-        .collect()
 }